Switch dtype to bfloat16 for efficiency in training config

+++ config/train_predictor_config.py
@@ -56,8 +56,8 @@ learning_rate = 3e-4
 
 # System settings (optimized for RunPod)
 backend = "nccl"
-# dtype = "bfloat16"  # Use bfloat16 for efficiency if available
-dtype = "float32"
+dtype = "bfloat16"  # Use bfloat16 for efficiency if available
+# dtype = "float32"
 compile = True
 keep_alive = False  # Auto-stop by default
 check_nans = False  # Check for NaNs in cloud training
